Tarefa 8 - Refactor: Posicao, Jogada e Jogador

+++ TI.Desempenho.Esportivo/Model/JogadaModel.cs
@@ -9,7 +9,7 @@ namespace TI.Desempenho.Esportivo.Model
     public class JogadaModel
     {
         public String nomJogada { get; set; }
-        public int pontosJogada { get; set; }
+        public float pontosJogada { get; set; }
         public float multiplicador { get; set; }
     }
 }
